box refactor; make box objects
