+++ pytc_gui/__init__.py
@@ -1,13 +1,13 @@
 __author__ = "Hiranmayi Duvvuri"
 __date__ = "2017-01-08"
-__all__ = ["sliders", "exp_frames", "slider_popup"]
+#__all__ = ["sliders", "exp_frames", "slider_popup"]
 
-from . import sliders
-from . import exp_frames
-from . import slider_popup
+#from . import sliders
+#from . import exp_frames
+#from . import slider_popup
 
-from .main_window import *
-from .visual import *
-from .sliders import *
-from .exp_frames import *
-from .slider_popup import *
+#from .main_window import *
+#from .visual import *
+#from .sliders import *
+#from .exp_frames import *
+#from .slider_popup import *
