+++ pytc_gui/boxes/exp_box/__init__.py
@@ -1,0 +1,8 @@
+__description__ = \
+"""
+Submodule holding the experimental box visual elements of the gui.
+"""
+__author__ = "Michael J. Harms"
+__date__ = "2017-06-01"
+
+from .exp_box import ExperimentBox
