+++ pytc_gui/boxes/exp_box/exp_frames/connectors_box.py
