+++ pytc_gui/boxes/exp_box/exp_frames/global_box.py
