+++ pytc_gui/boxes/exp_box/slider_popup/base.py
