+++ pytc_gui/boxes/exp_box/slider_popup/connector_popup.py
