+++ pytc_gui/boxes/exp_box/slider_popup/global_popup.py
