+++ pytc_gui/boxes/exp_box/sliders/add_global_connector.py
