+++ pytc_gui/boxes/exp_box/sliders/global_sliders.py
