+++ pytc_gui/boxes/plot_box.py
